+++ ni-measurement-ui-creator/examples/numeric_control_and_indicator.py
@@ -16,7 +16,6 @@ from ni_measurement_ui_creator.utils._helpers import (
     create_indicator_elements,
 )
 
-
 # Refer `SupportedDataType` class for supported `value_type`.
 
 # Any unique id will work.