+++ ni-measurement-ui-creator/src/ni_measurement_ui_creator/utils/_client.py
@@ -148,10 +148,12 @@ def get_measurement_service_stub(
 
     measurements = list(set([services.display_name for services in available_services]))
 
+    logger.info("")
     logger.info(UserMessage.AVAILABLE_MEASUREMENTS)
     for serial_num, services in enumerate(measurements):
         logger.info(f"{serial_num + 1}. {services}")
 
+    logger.info("")
     selected_measurement = get_measurement_selection(total_measurements=len(measurements))
 
     measurement_service_class = get_measurement_service_class(